archival: supporting bits from archival_policy -> segment_reupload
