+++ src/v/cluster/archival/archival_policy.cc
@@ -15,13 +15,6 @@
 #include "cluster/archival/segment_reupload.h"
 #include "config/configuration.h"
 #include "model/fundamental.h"
-#include "storage/disk_log_impl.h"
-#include "storage/fs_utils.h"
-#include "storage/offset_to_filepos.h"
-#include "storage/parser.h"
-#include "storage/segment.h"
-#include "storage/segment_set.h"
-#include "storage/version.h"
 
 #include <seastar/core/iostream.hh>
 #include <seastar/core/lowres_clock.hh>
@@ -35,105 +28,6 @@ constexpr size_t compacted_segment_size_multiplier{3};
 namespace archival {
 
 using namespace std::chrono_literals;
-
-std::ostream& operator<<(std::ostream& s, const upload_candidate& c) {
-    vassert(
-      c.sources.empty() || c.remote_sources.empty(),
-      "The upload candidate could have only local or only remote source");
-    if (c.sources.empty() && c.remote_sources.empty()) {
-        s << "{empty}";
-        return s;
-    }
-
-    std::vector<ss::sstring> source_names;
-    source_names.reserve(std::max(c.sources.size(), c.remote_sources.size()));
-    if (c.remote_sources.empty()) {
-        std::transform(
-          c.sources.begin(),
-          c.sources.end(),
-          std::back_inserter(source_names),
-          [](const auto& src) { return src->filename(); });
-    } else if (c.sources.empty()) {
-        std::transform(
-          c.remote_sources.begin(),
-          c.remote_sources.end(),
-          std::back_inserter(source_names),
-          [](const auto& src) { return src().native(); });
-    }
-
-    fmt::print(
-      s,
-      "{{source segment offsets: {}, exposed_name: {}, starting_offset: {}, "
-      "file_offset: {}, content_length: {}, final_offset: {}, "
-      "final_file_offset: {}, term: {}, source names: {}}}",
-      c.sources.front()->offsets(),
-      c.exposed_name,
-      c.starting_offset,
-      c.file_offset,
-      c.content_length,
-      c.final_offset,
-      c.final_file_offset,
-      c.term,
-      source_names);
-    return s;
-}
-
-std::ostream& operator<<(std::ostream& os, candidate_creation_error err) {
-    os << "candidate creation error: ";
-    switch (err) {
-    case candidate_creation_error::no_segments_collected:
-        return os << "no segments collected";
-    case candidate_creation_error::begin_offset_seek_error:
-        return os << "failed to seek begin offset";
-    case candidate_creation_error::end_offset_seek_error:
-        return os << "failed to seek end offset";
-    case candidate_creation_error::offset_inside_batch:
-        return os << "offset inside batch";
-    case candidate_creation_error::upload_size_unchanged:
-        return os << "size of candidate unchanged";
-    case candidate_creation_error::cannot_replace_manifest_entry:
-        return os << "candidate cannot replace manifest entry";
-    case candidate_creation_error::no_segment_for_begin_offset:
-        return os << "no segment for begin offset";
-    case candidate_creation_error::missing_ntp_config:
-        return os << "missing config for NTP";
-    case candidate_creation_error::failed_to_get_file_range:
-        return os << "failed to get file range for candidate";
-    case candidate_creation_error::zero_content_length:
-        return os << "candidate has no content";
-    case candidate_creation_error::concurrency_error:
-        return os << "collected segments are modified concurrently";
-    }
-}
-
-ss::log_level log_level_for_error(const candidate_creation_error& error) {
-    switch (error) {
-    case candidate_creation_error::no_segments_collected:
-    case candidate_creation_error::begin_offset_seek_error:
-    case candidate_creation_error::end_offset_seek_error:
-    case candidate_creation_error::upload_size_unchanged:
-    case candidate_creation_error::cannot_replace_manifest_entry:
-    case candidate_creation_error::no_segment_for_begin_offset:
-    case candidate_creation_error::failed_to_get_file_range:
-    case candidate_creation_error::zero_content_length:
-    case candidate_creation_error::concurrency_error:
-        return ss::log_level::debug;
-    case candidate_creation_error::offset_inside_batch:
-    case candidate_creation_error::missing_ntp_config:
-        return ss::log_level::warn;
-    }
-}
-
-std::ostream&
-operator<<(std::ostream& os, const skip_offset_range& skip_range) {
-    fmt::print(
-      os,
-      "skip_offset_range{{begin: {}, end: {}, error: {}}}",
-      skip_range.begin_offset,
-      skip_range.end_offset,
-      skip_range.reason);
-    return os;
-}
 
 archival_policy::archival_policy(
   model::ntp ntp, std::optional<segment_time_limit> limit)
